# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "KubePi Restful API",
        "termsOfService": "https://github.com/1Panel-dev/KubePi",
        "contact": {
            "name": "Fit2cloud Support",
            "url": "https://www.fit2cloud.com",
            "email": "support@fit2cloud.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/kubepi/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.UserCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.User"
                        }
                    }
                }
            }
        },
        "/users/search": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search users by Condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Search users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Page"
                        }
                    }
                }
            }
        },
        "/users/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.User"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update user by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.User"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete user by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete user by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Page": {
            "type": "object",
            "properties": {
                "items": {},
                "total": {
                    "type": "integer"
                }
            }
        },
        "docs.Authenticate": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "docs.Mfa": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                }
            }
        },
        "docs.UserCreate": {
            "type": "object",
            "properties": {
                "authenticate": {
                    "$ref": "#/definitions/docs.Authenticate"
                },
                "email": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "mfa": {
                    "$ref": "#/definitions/docs.Mfa"
                },
                "name": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_1Panel-dev_KubePi_internal_model_v1_user.Authenticate": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_user.Mfa": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "github_com_1Panel-dev_KubePi_internal_model_v1_user.User": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "authenticate": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.Authenticate"
                },
                "builtIn": {
                    "type": "boolean"
                },
                "createAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "mfa": {
                    "$ref": "#/definitions/github_com_1Panel-dev_KubePi_internal_model_v1_user.Mfa"
                },
                "name": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updateAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}